refactor(ui): decompose renderer.go and browser.go (<500 LOC each)
Part A — internal/ui/renderer.go (908→409 LOC):
- renderer_spinner.go (149 LOC): spinner lifecycle methods
- renderer_metrics.go (376 LOC): metrics rendering and logging methods
- renderer.go keeps core types, constructor, sub-renderers, markdown

Part B — internal/ui/tui/browser.go (858→374 LOC):
- browser_viewport.go (296 LOC): viewport, rendering, footer
- browser_search.go (48 LOC): search match calculation, highlighting
- browser_pinning.go (172 LOC): pin management, rollback
- browser.go keeps core model, message dispatch, entry points

Pure structural change — no semantic modifications. All tests pass
with -race, go vet clean, verify_architecture shows zero new edges.
