+++ internal/ui/tui/browser.go
@@ -8,7 +8,6 @@ import (
 	"fmt"
 	"log"
 	"os"
-	"regexp"
 	"strings"
 	"time"
 
@@ -150,13 +149,6 @@ func (m *rootBrowserModel) handleWatcherError(err error, ok bool) tea.Msg {
 	}
 	log.Printf("history file watcher error: %v", err)
 	return nil
-}
-
-func (m *rootBrowserModel) updateViewportHeight() {
-	if !m.ready {
-		return
-	}
-	m.viewport.Height = m.height - m.calculateFooterHeight()
 }
 
 // Update handles incoming messages and updates the model state.
@@ -358,63 +350,6 @@ func (m *rootBrowserModel) handleFileChangedMsg(msg fileChangedMsg) (tea.Model, 
 	)
 }
 
-func (m *rootBrowserModel) handleViewportUpdate(msg tea.Msg) (tea.Model, tea.Cmd) {
-	var (
-		cmd  tea.Cmd
-		cmds []tea.Cmd
-	)
-
-	// Forward messages to the viewport
-	m.viewport, cmd = m.viewport.Update(msg)
-	cmds = append(cmds, cmd)
-
-	// Infinite pagination trigger (Older history from archive)
-	if m.viewport.YOffset == 0 && !m.isLoading && m.cursor != "" && m.cursor != "EOF" && m.ready && !m.isSearching {
-		m.isLoading = true
-		m.updateViewportContent()
-		return m, tea.Batch(append(cmds, fetchHistoryCmd(m.provider, m.cursor))...)
-	}
-
-	return m, tea.Batch(cmds...)
-}
-
-func (m *rootBrowserModel) updateViewportContent() {
-	// 1. Invalidate cache if needed
-	if m.lastWidth != m.width || m.lastQuery != m.currentQuery {
-		m.cachedThoughts = make(map[string]string)
-		m.lastWidth = m.width
-		m.lastQuery = m.currentQuery
-	}
-
-	// 2. Populate cache for all history turns
-	for _, dto := range m.history {
-		if _, ok := m.cachedThoughts[dto.ID]; !ok && m.showThoughts && dto.ThoughtProcess != "" {
-			m.cachedThoughts[dto.ID] = m.preRenderThought(dto)
-		}
-	}
-
-	// 3. Render and update state
-	rendered, offsets := m.renderHistory()
-	m.turnOffsets = offsets
-	m.recalculateSearchMatches(rendered)
-	m.viewport.SetContent(rendered)
-}
-
-func (m *rootBrowserModel) preRenderThought(dto ports.HistoryViewDTO) string {
-	thoughtText := "💭 [THOUGHTS]\n" + dto.ThoughtProcess
-	if m.currentQuery != "" {
-		thoughtText = m.highlightMatches(thoughtText, m.currentQuery)
-	}
-
-	prefixLen := 2 // standard prefix "  " or "> "
-	maxWidth := m.width - prefixLen
-	if maxWidth < 20 {
-		maxWidth = 20
-	}
-
-	return thoughtStyle.Width(maxWidth).Render(thoughtText)
-}
-
 // View renders the current state of the model.
 func (m *rootBrowserModel) View() string {
 	if m.err != nil {
@@ -436,423 +371,4 @@ func (m *rootBrowserModel) View() string {
 	}
 
 	return sb.String()
-}
-
-func (m *rootBrowserModel) renderEmptyState() string {
-	if m.isLoading {
-		return "Loading history..."
-	}
-	if m.cursor == "EOF" {
-		return "No history found."
-	}
-	return ""
-}
-
-func (m *rootBrowserModel) renderHistoryStatus(sb *strings.Builder) {
-	if m.isLoading {
-		sb.WriteString(thoughtStyle.Render("Loading more messages...") + "\n\n")
-	} else if m.cursor == "EOF" && len(m.history) > 0 {
-		sb.WriteString(archivedStyle.Render("─── Start of History ───") + "\n\n")
-	}
-}
-
-func (m *rootBrowserModel) renderTurn(sb *strings.Builder, i int, dto ports.HistoryViewDTO) {
-	prefix := "  "
-	if i == m.selectedTurn {
-		prefix = "> "
-	}
-
-	sb.WriteString(m.renderTurnHeader(dto, i == m.selectedTurn))
-
-	if m.showThoughts && dto.ThoughtProcess != "" {
-		sb.WriteString(m.renderThoughts(dto, prefix))
-	}
-
-	if len(dto.ToolCalls) > 0 {
-		sb.WriteString(m.renderToolCalls(dto, prefix))
-	}
-
-	sb.WriteString(m.renderContent(dto, prefix))
-}
-
-func (m *rootBrowserModel) renderHistory() (string, []int) {
-	if len(m.history) == 0 {
-		return m.renderEmptyState(), nil
-	}
-
-	turnOffsets := make([]int, 0, len(m.history))
-	var sb strings.Builder
-
-	m.renderHistoryStatus(&sb)
-
-	for i, dto := range m.history {
-		turnOffsets = append(turnOffsets, strings.Count(sb.String(), "\n"))
-
-		m.renderTurn(&sb, i, dto)
-
-		if i < len(m.history)-1 {
-			sb.WriteString(m.renderSeparator())
-		}
-	}
-
-	return sb.String(), turnOffsets
-}
-
-func (m *rootBrowserModel) renderTurnHeader(dto ports.HistoryViewDTO, isSelected bool) string {
-	prefix := "  "
-	if isSelected {
-		prefix = "> "
-	}
-
-	roleLabel := m.getRoleLabel(dto)
-	turnStr := m.getTurnLabelSuffix(dto)
-	styledLabel := m.getStyledRoleLabel(dto.Role, roleLabel, turnStr)
-
-	if dto.IsArchived {
-		styledLabel += archivedStyle.Render(" (archived)")
-	}
-	if dto.IsPinned {
-		styledLabel += lipgloss.NewStyle().Foreground(lipgloss.Color("208")).Bold(true).Render(" [PINNED]")
-	}
-
-	return prefix + styledLabel + "\n"
-}
-
-func (m *rootBrowserModel) getRoleLabel(dto ports.HistoryViewDTO) string {
-	roleLabel := strings.ToUpper(dto.Role)
-	if dto.Role == "assistant" {
-		roleLabel = "MODEL"
-	}
-	return roleLabel
-}
-
-func (m *rootBrowserModel) getTurnLabelSuffix(dto ports.HistoryViewDTO) string {
-	if !dto.IsArchived {
-		turnIdx := m.getTurnIndex(dto)
-		if turnIdx >= 0 {
-			return fmt.Sprintf(" - %d", turnIdx+1)
-		}
-	}
-	return ""
-}
-
-func (m *rootBrowserModel) getStyledRoleLabel(role, label, suffix string) string {
-	switch role {
-	case "user":
-		return userStyle.Render(fmt.Sprintf("[%s]%s", label, suffix))
-	case "assistant", "model":
-		return modelStyle.Render(fmt.Sprintf("[%s]%s", label, suffix))
-	default:
-		return lipgloss.NewStyle().Bold(true).Render(fmt.Sprintf("[%s]%s", label, suffix))
-	}
-}
-
-func (m *rootBrowserModel) renderThoughts(dto ports.HistoryViewDTO, prefix string) string {
-	var wrappedText string
-
-	// Pure read from cache
-	if cached, ok := m.cachedThoughts[dto.ID]; ok {
-		wrappedText = cached
-	} else {
-		// Computation only, no storage in m.cachedThoughts to keep this pure
-		thoughtText := "💭 [THOUGHTS]\n" + dto.ThoughtProcess
-		if m.currentQuery != "" {
-			thoughtText = m.highlightMatches(thoughtText, m.currentQuery)
-		}
-
-		maxWidth := m.width - len(prefix)
-		if maxWidth < 20 {
-			maxWidth = 20
-		}
-
-		wrappedText = thoughtStyle.Width(maxWidth).Render(thoughtText)
-	}
-
-	// Apply the dynamic prefix
-	var sb strings.Builder
-	lines := strings.Split(wrappedText, "\n")
-	for _, line := range lines {
-		sb.WriteString(prefix)
-		sb.WriteString(line)
-		sb.WriteString("\n")
-	}
-	sb.WriteString("\n")
-
-	return sb.String()
-}
-
-func (m *rootBrowserModel) renderToolCalls(dto ports.HistoryViewDTO, prefix string) string {
-	var sb strings.Builder
-	for _, tool := range dto.ToolCalls {
-		sb.WriteString(prefix)
-		sb.WriteString(toolStyle.Render(fmt.Sprintf("  🔧 Executing tool: %s", tool)))
-		sb.WriteString("\n")
-	}
-	sb.WriteString("\n")
-	return sb.String()
-}
-
-func (m *rootBrowserModel) renderContent(dto ports.HistoryViewDTO, prefix string) string {
-	content := dto.ContentPreview
-	if m.currentQuery != "" {
-		content = m.highlightMatches(content, m.currentQuery)
-	}
-
-	var sb strings.Builder
-	lines := strings.Split(content, "\n")
-	for j, line := range lines {
-		sb.WriteString(prefix)
-		sb.WriteString(line)
-		if j < len(lines)-1 {
-			sb.WriteString("\n")
-		}
-	}
-	return sb.String()
-}
-
-func (m *rootBrowserModel) renderSeparator() string {
-	return "\n\n" + archivedStyle.Render(strings.Repeat("─", m.width/2)) + "\n\n"
-}
-
-func (m *rootBrowserModel) recalculateSearchMatches(rendered string) {
-	m.matches = []int{}
-	if m.currentQuery != "" {
-		re, err := regexp.Compile("(?i)" + regexp.QuoteMeta(m.currentQuery))
-		if err == nil {
-			lines := strings.Split(rendered, "\n")
-			for i, line := range lines {
-				if re.MatchString(line) {
-					m.matches = append(m.matches, i)
-				}
-			}
-		}
-	}
-
-	if len(m.matches) > 0 {
-		if m.currentMatch >= len(m.matches) {
-			m.currentMatch = len(m.matches) - 1
-		}
-	} else {
-		m.currentMatch = 0
-	}
-}
-
-func (m *rootBrowserModel) highlightMatches(text, query string) string {
-	if query == "" {
-		return text
-	}
-
-	// Case-insensitive regex for the query
-	re, err := regexp.Compile("(?i)" + regexp.QuoteMeta(query))
-	if err != nil {
-		return text
-	}
-
-	return re.ReplaceAllStringFunc(text, func(match string) string {
-		return highlightStyle.Render(match)
-	})
-}
-
-func (m *rootBrowserModel) calculateFooterHeight() int {
-	if m.isSearching {
-		return 1
-	}
-
-	activeTurns, pinnedTurns := m.getPinningMetrics()
-
-	if pinnedTurns > 5 || (activeTurns > 0 && float64(pinnedTurns)/float64(activeTurns) > 0.5) {
-		return 2
-	}
-	return 1
-}
-
-func (m *rootBrowserModel) renderFooter() string {
-	var sb strings.Builder
-
-	activeTurns, pinnedTurns := m.getPinningMetrics()
-
-	if pinnedTurns > 5 || (activeTurns > 0 && float64(pinnedTurns)/float64(activeTurns) > 0.5) {
-		sb.WriteString(warningStyle.Render("⚠️ High Pinning Pressure: Auto-summarization may fail."))
-		sb.WriteString("\n")
-	}
-
-	thoughtsStatus := "ON"
-	if !m.showThoughts {
-		thoughtsStatus = "OFF"
-	}
-	fmt.Fprintf(&sb, "↑/↓: Scroll • Space: Thoughts [%s] • /: Search • j/k: Select • p: Pin • r: Rollback • q: Quit", thoughtsStatus)
-
-	if m.currentQuery != "" {
-		matchInfo := ""
-		if len(m.matches) > 0 {
-			matchInfo = fmt.Sprintf(" (%d/%d matches)", m.currentMatch+1, len(m.matches))
-		} else {
-			matchInfo = " (no matches)"
-		}
-		fmt.Fprintf(&sb, " • Query: %q%s", m.currentQuery, matchInfo)
-	}
-	if m.isLoading {
-		sb.WriteString(" • LOADING...")
-	}
-	return footerStyle.Render(sb.String())
-}
-
-func fetchHistoryCmd(provider ports.UnifiedHistoryProvider, cursor string) tea.Cmd {
-	return func() tea.Msg {
-		dtos, nextCursor, err := provider.GetHistoryStream(context.Background(), 20, cursor)
-		if err == nil && len(dtos) == 0 && cursor != "" {
-			nextCursor = "EOF"
-		}
-		return historyLoadedMsg{
-			dtos:       dtos,
-			nextCursor: nextCursor,
-			err:        err,
-		}
-	}
-}
-
-func (m *rootBrowserModel) syncViewportToSelectedTurn() {
-	if m.selectedTurn >= 0 && m.selectedTurn < len(m.turnOffsets) {
-		targetLine := m.turnOffsets[m.selectedTurn]
-		if targetLine < m.viewport.YOffset {
-			m.viewport.SetYOffset(targetLine)
-		}
-		if targetLine >= m.viewport.YOffset+m.viewport.Height {
-			m.viewport.SetYOffset(targetLine - m.viewport.Height + 1)
-		}
-	}
-}
-
-func (m *rootBrowserModel) getSystemOffset() int {
-	for _, dto := range m.history {
-		if dto.OriginalIndex == 0 && dto.Role == "system" {
-			return 1
-		}
-	}
-	return 0
-}
-
-// getTurnIndex returns the 0-based turn number (pair of msgs) for a given DTO.
-func (m *rootBrowserModel) getTurnIndex(dto ports.HistoryViewDTO) int {
-	offset := m.getSystemOffset()
-	if dto.OriginalIndex < offset && dto.Role == "system" {
-		return -1
-	}
-	return (dto.OriginalIndex - offset) / 2
-}
-
-// getTurnStartOriginalIndex returns the index of the first message in a turn (usually the 'user' msg).
-func (m *rootBrowserModel) getTurnStartOriginalIndex(dto ports.HistoryViewDTO) int {
-	offset := m.getSystemOffset()
-	msgOffset := dto.OriginalIndex - offset
-	if msgOffset < 0 {
-		return dto.OriginalIndex
-	}
-	return (msgOffset & ^1) + offset
-}
-
-func (m *rootBrowserModel) getPinningMetrics() (activeTurns int, pinnedTurns int) {
-	lastTurnIdx := -1
-	for _, dto := range m.history {
-		if dto.IsArchived {
-			continue
-		}
-		turnIdx := m.getTurnIndex(dto)
-		if turnIdx < 0 {
-			continue
-		}
-		if turnIdx != lastTurnIdx {
-			activeTurns++
-			if dto.IsPinned {
-				pinnedTurns++
-			}
-			lastTurnIdx = turnIdx
-		}
-	}
-	return activeTurns, pinnedTurns
-}
-
-func (m *rootBrowserModel) togglePin() {
-	dto, turnIdx, ok := m.getTurnForPinning()
-	if !ok {
-		return
-	}
-
-	// Toggle pin state
-	err := m.cmdService.SetPinned(context.Background(), turnIdx, !dto.IsPinned)
-	if err != nil {
-		m.err = err
-		return
-	}
-
-	m.updateLocalPinState(dto, !dto.IsPinned)
-	m.lastMutationTime = time.Now()
-	m.updateViewportContent()
-	m.updateViewportHeight()
-}
-
-func (m *rootBrowserModel) getTurnForPinning() (ports.HistoryViewDTO, int, bool) {
-	if m.selectedTurn == -1 || m.selectedTurn >= len(m.history) {
-		return ports.HistoryViewDTO{}, 0, false
-	}
-
-	dto := m.history[m.selectedTurn]
-	if dto.IsArchived {
-		return ports.HistoryViewDTO{}, 0, false
-	}
-
-	turnIdx := m.getTurnIndex(dto)
-	if turnIdx < 0 {
-		return ports.HistoryViewDTO{}, 0, false
-	}
-	return dto, turnIdx, true
-}
-
-func (m *rootBrowserModel) updateLocalPinState(dto ports.HistoryViewDTO, newPinState bool) {
-	turnStartIdx := m.getTurnStartOriginalIndex(dto)
-	for idx := range m.history {
-		if !m.history[idx].IsArchived && m.getTurnStartOriginalIndex(m.history[idx]) == turnStartIdx {
-			m.history[idx].IsPinned = newPinState
-		}
-	}
-}
-
-func (m *rootBrowserModel) rollbackToSelected() tea.Cmd {
-	if m.selectedTurn == -1 || m.selectedTurn >= len(m.history) {
-		return nil
-	}
-
-	dto := m.history[m.selectedTurn]
-	if dto.IsArchived {
-		return nil
-	}
-
-	// We need total active messages to calculate how many turns to remove.
-	lastActiveIdx := -1
-	for i := len(m.history) - 1; i >= 0; i-- {
-		if !m.history[i].IsArchived {
-			lastActiveIdx = m.history[i].OriginalIndex
-			break
-		}
-	}
-	if lastActiveIdx == -1 {
-		return nil
-	}
-
-	targetStartIdx := m.getTurnStartOriginalIndex(dto)
-	turnsToRemove := (lastActiveIdx - targetStartIdx + 1 + 1) / 2
-
-	_, _, _, err := m.cmdService.RollbackTurns(context.Background(), turnsToRemove)
-	if err != nil {
-		m.err = err
-		return nil
-	}
-
-	m.lastMutationTime = time.Now()
-	// Full Refresh
-	m.history = nil
-	m.cursor = ""
-	m.selectedTurn = -1
-	m.isLoading = true
-	return fetchHistoryCmd(m.provider, "")
 }
